Refactor for readability
